Tweaks due to previous commit. Split wavefunc_gen to a new file.
